Conflicts:
	code/core/src/ir_builder.cpp

+++ code/runtime/include/include_gems/stdint.h
@@ -40,7 +40,7 @@
 
 #define UINT64_MAX (18446744073709551615ULL)
 
-typedef int intptr_t;
-typedef unsigned int uintptr_t;
+typedef long long intptr_t;
+typedef unsigned long long uintptr_t;
 
 #endif // ifndef __GUARD_INCLUDE_GEMS_STDINT_H
